Upgrade to servicemix-utils 1.5.0-SNAPSHOT
git-svn-id: https://svn.apache.org/repos/asf/servicemix/components/trunk@1085746 13f79535-47bb-0310-9956-ffa450edef68

+++ shared-libraries/servicemix-common/src/main/java/org/apache/servicemix/common/osgi/Configuration.java
@@ -88,7 +88,7 @@ public class Configuration {
      * Set the keep alive time on the executor
      * @see org.apache.servicemix.executors.impl.ExecutorConfig#setKeepAliveTime(long) 
      */
-    public void setKeepAliveTime(int time) {
+    public void setKeepAliveTime(long time) {
         logger.debug("Setting thread keep-alive time: {}", time);
         executorConfig.setKeepAliveTime(time);
     }
